[RF] Avoid compile time dependency of RooFit on RDataFrame
For better modularity in the future, the major end-user ROOT components
like RDataFrame, RooFit, and TMVA should not depend on each other.

There was a dependency of RooFit on RDataFrame because of the
RooAbsDataHelper, but it is actually easy to avoid this with forward
declarations.

This should have no impact on the user, because if they use RDF, they
probably include it already if they're not using the interpreter.
