+++ roofit/CMakeLists.txt
@@ -18,7 +18,6 @@ if(mathmore)
 endif()
 add_subdirectory(roostats)
 add_subdirectory(histfactory)
-add_subdirectory(RDataFrameHelpers)
 add_subdirectory(jsoninterface)
 add_subdirectory(hs3)
 if(roofit_legacy_eval_backend AND NOT MSVC)
