--- roofit/RDataFrameHelpers/inc/LinkDef.h
@@ -1,6 +1,0 @@
-#ifdef __CLING__
-
-#pragma link off all globals;
-#pragma link off all classes;
-
-#endif
